// src/api/types/RulesengineFlag.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

import * as Schematic from "../index";

export interface RulesengineFlag {
    accountId: string;
    defaultValue: boolean;
    environmentId: string;
    id: string;
    key: string;
    rules: Schematic.RulesengineRule[];
}